lambda_deployment/requirements.txt
boto3
strands-agents-tools
strands-agents-builder
pydantic>=2.0.0
anthropic
